% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations-class.R
\docType{class}
\name{Annotations-class}
\alias{Annotations}
\alias{Annotations-class}
\title{An annotations table}
\description{
This table (in TSV format) must at minimum contain SHA256 hash values in column \emph{SHA256_value} and file paths relative to a single documents root directory that point to XML documents in column \emph{relative_file_paths}.
}
\section{Slots}{

\describe{
\item{\code{DT}}{A \code{\link[data.table]{data.table}} containing annotations, with rows as observations and columns as variables.}
}}

