% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArcCategories-class.R
\docType{class}
\name{ArcCategories-class}
\alias{ArcCategories}
\alias{ArcCategories-class}
\title{A set of arc category sets}
\description{
A set of arc category sets
}
\section{Slots}{

\describe{
\item{\code{DT_LST}}{A list of \code{\link[data.table]{data.table}} objects, one for each arc category, with rows describing arc labels and the first column being the arc label (i.e., its name) and the second a description. This table in essence descibes an enumeration. It is necessary that the arc labels do not contain whitespace, and it is advisable to use very brief arc labels in order to keep space requirements down.}

\item{\code{PRIOR_PROBABILITY_DVEC}}{Ordered by \code{DT_LST}, for each arc category the prior probability of observing an arc labeled by any of the arc labels in this arc category. It is assumed to be the uniform probability.}
}}
\concept{
structural
}

