% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchFEAT.R
\name{batchFEAT}
\alias{batchFEAT}
\title{Command-line FEAT function to perform analysis in batch for a given experiment.}
\usage{
batchFEAT(biom_table, mapping_file, FMT_pairs, input_params, output_dir)
}
\arguments{
\item{biom_table}{BIOM table}

\item{mapping_file}{Mapping File}

\item{FMT_pairs}{tsv indicating FMT pairs (see example)}

\item{input_params}{input parameters in tsv}

\item{output_dir}{output directory to leave output files}
}
\value{
many files containing QC and summary statistics, as well as FMT visualizations
}
\description{
Command-line FEAT function to perform analysis in batch for a given experiment.
}

