% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{collapse_taxonomy}
\alias{collapse_taxonomy}
\title{Define a function that will collapse a table by shared taxonomy and average the mean/median OTU count & differences for OTUs that share taxonomy}
\usage{
collapse_taxonomy(table)
}
\value{
table collapsed by taxonomy
}
\description{
Define a function that will collapse a table by shared taxonomy and average the mean/median OTU count & differences for OTUs that share taxonomy
}

