% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignMatrix-class.R
\docType{class}
\name{DesignMatrix-class}
\alias{DesignMatrix}
\alias{DesignMatrix-class}
\title{A design matrix}
\description{
A design matrix
}
\section{Slots}{

\describe{
\item{\code{DMAT}}{A design matrix with feature scores, with objects as rows indexed by their \emph{object_ID} and named columns as features. May contain a column named the empty string, with all-zero values (\code{DUMMY_OBJECT_FEATURES_DVEC} in the code). This is the case when from one or more objects no structural features could be extracted given the feature representation parameters.}

\item{\code{LOWER_RANK_APPROXIMATION_DMAT}}{A lower rank approximation of \code{DMAT}, based on \code{SVD}. See: \code{\link{FeatureRepresentation}}.}

\item{\code{SVD}}{A \emph{Singular Value Decomposition} (SVD) of \code{DMAT} as returned by \code{\link[base]{svd}}. Will be \code{NULL} if SVD has not been applied (\code{\link{FeatureRepresentation}}).}
}}

