% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features-generic.R, R/extract_features-methods.R
\docType{methods}
\name{extract_features}
\alias{extract_features}
\alias{extract_features,SrfFeatureRepresentation,FeaturesExtractionParameters,integer,character,character-method}
\alias{extract_features,StrctFeatureRepresentation,FeaturesExtractionParameters,integer,character,character-method}
\title{To extract features of an object}
\usage{
extract_features(object, FEATURES_EXTRACTION_PARAMETERS, OBJECT_ID_I,
  OBJECT_OUTPUT_DIR_PATH_STR, RELATIVE_FILE_PATHS_CVEC)


  \S4method{extract_features}{SrfFeatureRepresentation,FeaturesExtractionParameters,integer,character,character}(object,
  FEATURES_EXTRACTION_PARAMETERS, OBJECT_ID_I, OBJECT_OUTPUT_DIR_PATH_STR,
  RELATIVE_FILE_PATHS_CVEC)


  \S4method{extract_features}{StrctFeatureRepresentation,FeaturesExtractionParameters,integer,character,character}(object,
  FEATURES_EXTRACTION_PARAMETERS, OBJECT_ID_I, OBJECT_OUTPUT_DIR_PATH_STR,
  RELATIVE_FILE_PATHS_CVEC)
}
\arguments{
\item{object}{See: \code{\link{StrctFeatureRepresentation}} and \code{\link{SrfFeatureRepresentation}}.}

\item{FEATURES_EXTRACTION_PARAMETERS}{See: \code{\link{FeaturesExtractionParameters}}.}

\item{OBJECT_ID_I}{The \emph{object_ID} of the object.}

\item{OBJECT_OUTPUT_DIR_PATH_STR}{The directory path for output about the object.}

\item{RELATIVE_FILE_PATHS_CVEC}{The file paths relative to the documents root directory that collectively represent the object.}
}
\value{
\code{OBJECT_FEATURES_DVEC} Object features.
}
\description{
Extracts features of an object given a feature representation.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SrfFeatureRepresentation,FEATURES_EXTRACTION_PARAMETERS = FeaturesExtractionParameters,OBJECT_ID_I = integer,OBJECT_OUTPUT_DIR_PATH_STR = character,RELATIVE_FILE_PATHS_CVEC = character}: See: \code{\link{SrfFeatureRepresentation}}

\item \code{object = StrctFeatureRepresentation,FEATURES_EXTRACTION_PARAMETERS = FeaturesExtractionParameters,OBJECT_ID_I = integer,OBJECT_OUTPUT_DIR_PATH_STR = character,RELATIVE_FILE_PATHS_CVEC = character}: See: \code{\link{StrctFeatureRepresentation}}
}}
\concept{
surface

structural
}
\keyword{methods}

