% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureRepresentation-class.R
\docType{class}
\name{FeatureRepresentation-class}
\alias{FeatureRepresentation}
\alias{FeatureRepresentation-class}
\title{A virtual class for feature representations}
\description{
A virtual class for feature representations
}

