% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturesExtractionParameters-class.R
\docType{class}
\name{FeaturesExtractionParameters-class}
\alias{FeaturesExtractionParameters}
\alias{FeaturesExtractionParameters-class}
\title{Parameters of the feature extraction job}
\description{
Parameters of the feature extraction job
}
\section{Slots}{

\describe{
\item{\code{ANNOTATIONS_TABLE_FILE_PATH}}{The file path to the annotations table. See: \code{\link{Annotations}}, \code{\link{metadata_write_annotations_table_template}}, \code{\link{metadata_verify_annotation_table_consistency}}, \code{\link{metadata_read_annotations_table}}.}

\item{\code{DOCUMENTS_DIR_PATH}}{The root directory path containing the XML documents representing observations. To this the \code{relative_file_path} column in annotation tables must be relative.}

\item{\code{JOB_NAME_STR}}{A name for the particular feature extraction job.}

\item{\code{OUTPUT_DIR_PATH}}{The root directory path for any filesystem output.}
}}

