% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesystemPath-class.R
\docType{class}
\name{FilesystemPath-class}
\alias{FilesystemPath}
\alias{FilesystemPath-class}
\title{A string representing a filesystem path}
\description{
A string representing a filesystem path
}

