% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{filter_by_metric}
\alias{filter_by_metric}
\title{Define a function that filters tables by a minimum difference between means/medians}
\usage{
filter_by_metric(table, min_diff, difference_labels)
}
\value{
table filtered for a minimum difference between categories
}
\description{
Define a function that filters tables by a minimum difference between means/medians
}

