% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{filter_function}
\alias{filter_function}
\title{Define a function to produce tables populated only by OTUs that have non-zero counts in a given fraction of all of the samples for that condition}
\usage{
filter_function(table, min_fraction, condition)
}
\value{
tables populated only by OTUs that have non-zero counts in a given fraction of all of the samples for that condition
}
\description{
Define a function to produce tables populated only by OTUs that have non-zero counts in a given fraction of all of the samples for that condition
}

