% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{get_condition_only_table_function}
\alias{get_condition_only_table_function}
\title{Function to get tables that are specific for each given condition}
\usage{
get_condition_only_table_function(table, category, value)
}
\value{
data table with only one condition represented
}
\description{
Function to get tables that are specific for each given condition
}

