% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchFEAT}
\alias{launchFEAT}
\title{Launch shiny app to analyze FMT efficacy}
\usage{
launchFEAT()
}
\value{
nothing - will launch Shiny app in viewer
}
\description{
Launch shiny app to analyze FMT efficacy
}

