% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{means_of_numbers_only}
\alias{means_of_numbers_only}
\title{Define a function that will take a vector and return the mean of the vector only if it is a vector of numbers but returns the first instance of the vector when it is not a number}
\usage{
means_of_numbers_only(x)
}
\value{
an average of only the numbers of a given vector
}
\description{
Define a function that will take a vector and return the mean of the vector only if it is a vector of numbers but returns the first instance of the vector when it is not a number
}

