% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_read_annotations_table}
\alias{metadata_read_annotations_table}
\title{To read an annotations table template}
\usage{
metadata_read_annotations_table(ANNOTATIONS_TABLE_FILE_PATH_STR,
  KEPT_ANNOTATIONS_DT_COLUMN_NAMES_CVEC, ...)
}
\arguments{
\item{ANNOTATIONS_TABLE_FILE_PATH_STR}{The annotations table file path.}

\item{KEPT_ANNOTATIONS_DT_COLUMN_NAMES_CVEC}{The names of the columns of the annotations table to preserve in the return value.}

\item{...}{Extra arguments passed on to \code{\link[data.table]{fread}}.}
}
\value{
\code{ANNOTATIONS} See: \link{Annotations}.
}
\description{
This template should be manually annotated for feature extraction and supervised learning purposes.
}

