% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_verify_annotation_table_consistency}
\alias{metadata_verify_annotation_table_consistency}
\title{To verify an annotations table template}
\usage{
metadata_verify_annotation_table_consistency(ANNOTATIONS_DT,
  DOCUMENTS_DIR_PATH_STR)
}
\arguments{
\item{ANNOTATIONS_DT}{The annotations \code{\link[data.table]{data.table}}. See: \link{Annotations}.}

\item{DOCUMENTS_DIR_PATH_STR}{The root documents directory path.}
}
\value{
\emph{Signals an error if not verified.}
}
\description{
Performs a consistency check on the annotations table by comparing the SHA256 hash values in column \emph{SHA256_value} the annotations table with actual SHA256 hash alue of the respective XML document files under \code{DOCUMENTS_DIR_PATH_STR}.
}

