% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_write_annotations_table_template}
\alias{metadata_write_annotations_table_template}
\title{To write an annotations table template}
\usage{
metadata_write_annotations_table_template(ANNOTATIONS_TABLE_FILE_PATH_STR,
  DOCUMENTS_DIR_PATH_STR)
}
\arguments{
\item{ANNOTATIONS_TABLE_FILE_PATH_STR}{The annotations table file path.}

\item{DOCUMENTS_DIR_PATH_STR}{The root documents directory path.}
}
\description{
This template should be manually annotated for feature extraction and supervised learning purposes.
}

