% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{nonzero_fraction_getter}
\alias{nonzero_fraction_getter}
\title{function that returns a table of OTUs and the fraction of samples in which they have non-zero values}
\usage{
nonzero_fraction_getter(table)
}
\arguments{
\item{table}{OTU table}
}
\value{
table of OTUs and the fraction of samples in which they have non-zero values
}
\description{
function that returns a table of OTUs and the fraction of samples in which they have non-zero values
}

