% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{objects_process_collection}
\alias{objects_process_collection}
\title{To process all objects in a collection with one or more feature representations}
\usage{
objects_process_collection(DT, FEATURES_EXTRACTION_PARAMETERS,
  FEATURE_REPRESENTATIONS_LST)
}
\arguments{
\item{DT}{The annotations table as a \code{\link[data.table]{data.table}}. See: \code{\link{Annotations}}.}

\item{FEATURES_EXTRACTION_PARAMETERS}{See: \code{\link{FeaturesExtractionParameters}}.}

\item{FEATURE_REPRESENTATIONS_LST}{See: \code{\link{SrfFeatureRepresentation}}, \code{\link{StrctFeatureRepresentation}}.}
}
\value{
The \code{DESIGN_DMAT} or design matrix with feature scores, with objects as rows indexed by their \emph{object_ID} and named columns as features.
}
\description{
To process all objects in a collection with one or more feature representations
}
\keyword{file}

