% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{objects_read_or_process_with_feature_representation}
\alias{objects_read_or_process_with_feature_representation}
\title{To process objects according to a feature representation (performs filesystem input)}
\usage{
objects_read_or_process_with_feature_representation(FEATURE_REPRESENTATION, DT,
  FEATURES_EXTRACTION_PARAMETERS)
}
\arguments{
\item{FEATURE_REPRESENTATION}{See: \code{\link{FeatureRepresentation}}.}

\item{DT}{The annotations table as a \code{\link[data.table]{data.table}}. See: \code{\link{Annotations}}.}

\item{FEATURES_EXTRACTION_PARAMETERS}{See: \code{\link{FeaturesExtractionParameters}}.}
}
\value{
\code{DESIGN_DMAT} See: \code{\link{objects_process_with_feature_representation}}.
}
\description{
To process objects according to a feature representation (performs filesystem input)
}
\keyword{file}

