% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{percent}
\alias{percent}
\title{Make a percentage}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{decimal number}
}
\value{
a percentage from a decimal
}
\description{
Make a percentage
}

