% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{split_into_experiment}
\alias{split_into_experiment}
\title{Within Shiny environment, split a data table into an experiment-specific table given the input parameters}
\usage{
split_into_experiment(biom_table, mapping, compare, donor, recipient, post_fmt)
}
\value{
output table with only data from selected conditions
}
\description{
Within Shiny environment, split a data table into an experiment-specific table given the input parameters
}

