% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{split_into_experiment_batch}
\alias{split_into_experiment_batch}
\title{Split a data table into an experiment-specific table given the input parameters}
\usage{
split_into_experiment_batch(biom_table, mapping, compare, donor, recipient,
  post_fmt)
}
\arguments{
\item{biom_table}{OTU table}

\item{mapping}{mapping table}

\item{compare}{comparison category}

\item{donor}{donor condition}

\item{recipient}{recipient condition}

\item{post_fmt}{post-fmt condition}
}
\value{
output table with only data from selected conditions
}
\description{
Split a data table into an experiment-specific table given the input parameters
}

