% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrfFeatureRepresentation-class.R
\docType{class}
\name{SrfFeatureRepresentation-class}
\alias{SrfFeatureRepresentation}
\alias{SrfFeatureRepresentation-class}
\title{Surface ('SRF') feature representation}
\description{
Surface ('SRF') feature representation
}
\section{Slots}{

\describe{
\item{\code{NAME_STR}}{\emph{private} \code{'SRF'}}

\item{\code{N_I}}{The maximum value \eqn{N} for which \eqn{n}-grams will be extracted from object string sequences, with \eqn{n \in \{1,...,N\}}.}

\item{\code{OBJECT_SEPARATOR_STR}}{The separator used when collapsing strings representing single observations into object string sequences (e.g., texts).}

\item{\code{SRILM_NGRAMCOUNT_PARAMETERS_CVEC}}{Extra command line parameters to SRILM \emph{ngram-count}. This utility is run to extract features.}

\item{\code{SRILM_NGRAM_PARAMETERS_CVEC}}{Extra command line parameters to SRILM \emph{ngram}. This command is run to calculate feature scores.}

\item{\code{XPATH_STR}}{XPath used to extract the surface strings in the XML documents that represent observations. It ought to point to a text node.}
}}
\concept{
surface
}

