% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strct.R
\name{strct_read_arc_category_tables}
\alias{strct_read_arc_category_tables}
\title{To read arc category tables}
\usage{
strct_read_arc_category_tables(ARC_CATEGORIES_CVEC,
  ARC_CATEGORY_FILES_NAMES_REX_STR = "^arc_category_(\\\\w+).tsv$",
  ARC_CATEGORY_TABLES_DIR_PATH)
}
\arguments{
\item{ARC_CATEGORIES_CVEC}{The names used to identitify the arc categories, for which trails may be extracted by \code{\link{extract_features}}.}

\item{ARC_CATEGORY_FILES_NAMES_REX_STR}{A \emph{PCRE} regular expression that exclusively matches the arc category table files.}

\item{ARC_CATEGORY_TABLES_DIR_PATH}{A \code{\link{FilesystemPath}} to the directory containing an arc category table file for each arc category in \code{ARC_CATEGORIES_CVEC}.}
}
\value{
\code{ARC_CATEGORIES} See: \code{\link{ArcCategories}}.
}
\description{
To read arc category tables
}

