% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{table_reorder_first}
\alias{table_reorder_first}
\title{function to reorder tables with a given value first}
\usage{
table_reorder_first(table, value)
}
\value{
table with selected column moved to the front
}
\description{
function to reorder tables with a given value first
}

