% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{visualize_metrics}
\alias{visualize_metrics}
\title{Function to produce FMT metric visualizations}
\usage{
visualize_metrics(N_donor, N_recipient, N_postFMT, N_P_Don, N_P_Rec,
  N_unique_postfmt, N_shared_postfmt, title_text)
}
\arguments{
\item{N_donor}{number of otus in donor}

\item{N_recipient}{number of otus in recipient}

\item{N_postFMT}{number of otus in post-fmt}

\item{N_unique_postfmt}{number of post-fmt otus that are unique to post-fmt condition}

\item{N_shared_postfmt}{number of post-fmt otus that are shared between donor and recipient}

\item{title_text}{Text to generate title}

\item{N_P_Donoror}{number of post-fmt otus from donor}

\item{N_P_Recipientipient}{number of post-fmt otus from recipient}
}
\value{
a plot that depicts FMT efficacy
}
\description{
Function to produce FMT metric visualizations
}

