% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{normalize_and_filter}
\alias{normalize_and_filter}
\title{Within Shiny environment, normalize OTU counts for your OTU table, and impose a minimum filter}
\usage{
normalize_and_filter(biom_table, min_fraction)
}
\value{
output OTU table with relative abundances above the provided minimum
}
\description{
Within Shiny environment, normalize OTU counts for your OTU table, and impose a minimum filter
}

