% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{visualize_metrics}
\alias{visualize_metrics}
\title{Function to produce FMT metric visualizations}
\usage{
visualize_metrics(N_donor, N_recipient, N_postFMT, FMT_donor, FMT_recipient,
  N_unique_postfmt, N_shared_postfmt, title_text)
}
\value{
a plot that depicts FMT efficacy
}
\description{
Function to produce FMT metric visualizations
}

